+++ aegis_moveit_config/launch/move_group.launch.py
@@ -17,6 +17,15 @@ def str2bool(x: str) -> bool:
     return x.lower() in ("true")
 
 
+def deep_dict_update(d_target: dict, d_update: dict) -> dict:
+    for k, v in d_update.items():
+        if k in d_target and isinstance(d_target[k], dict) and isinstance(v, dict):
+            deep_dict_update(d_target[k], v)
+        else:
+            d_target[k] = v
+    return d_target
+
+
 class AegisPathsCfg:
     """Contains paths to the configuration files."""
 
@@ -59,6 +68,11 @@ class AegisPathsCfg:
     def load_joint_limits_cfg(self) -> dict:
         return load_yaml(self.description_cfg_pkg_name, "config/ur5e/joint_limits.yaml")
 
+    def load_planning_joint_limits_cfg(self) -> dict:
+        return load_yaml(
+            self.moveit_cfg_pkg_name, "config/move_group/planning_joint_limits.yaml"
+        )
+
     def load_octomap_updater_cfg(self) -> dict:
         return load_yaml(self.moveit_cfg_pkg_name, "config/octomap_updater.yaml")
 
@@ -77,8 +91,12 @@ def launch_setup(context: LaunchContext) -> list[Node]:
 
     mock_hardware_bool = str2bool(context.perform_substitution(mock_hardware))
 
+    # Planning joints limits
     robot_description_planning = {
-        "robot_description_planning": paths.load_joint_limits_cfg()
+        "robot_description_planning": deep_dict_update(
+            paths.load_joint_limits_cfg(),
+            paths.load_planning_joint_limits_cfg(),
+        )
     }
 
     # Planning Configuration
@@ -102,6 +120,10 @@ def launch_setup(context: LaunchContext) -> list[Node]:
         "trajectory_execution.allowed_execution_duration_scaling": 1.2,
         "trajectory_execution.allowed_goal_duration_margin": 0.5,
         "trajectory_execution.allowed_start_tolerance": 0.01,
+        # UR Driver is not compatible with the MoveIt's Trajectory Execution Monitoring (TEM)
+        # See # https://docs.universal-robots.com/Universal_Robots_ROS2_Documentation/doc/ur_robot_driver/ur_moveit_config/doc/index.html#id2
+        # Execution time monitoring can be incompatible with the scaled JTC
+        "trajectory_execution.execution_duration_monitoring": False,
     }
 
     planning_scene_monitor_parameters = {